Revert "MainActivity を ListFragment に移行 #2"
