+++ app/src/main/AndroidManifest.xml
@@ -16,12 +16,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Mybrary">
-        <activity android:name=".RecordPostActivity" />
+        <activity android:name=".RecordPostActivity"></activity>
         <activity android:name=".RecordActivity" />
         <activity android:name=".BookPostActivity" />
         <activity android:name=".ReadActivity" />
-        <activity android:name=".MainActivity"
-            android:exported="true">
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
