--- app/src/main/java/app/doggy/newmybrary/viewmodel/ListViewModel.kt
@@ -1,4 +1,0 @@
-package app.doggy.newmybrary.viewmodel
-
-class ListViewModel {
-}
