+++ app/src/main/res/values/strings.xml
@@ -31,7 +31,7 @@
     <string name="delete_toast_text_after">" deleted</string>
     <string name="delete_record_dialog_message">Are you sure you want to delete this record?</string>
     <string name="delete_record_toast_text">deleted</string>
-
+    
     <!--Item Cell-->
     <string name="current_page_text">Current p. </string>
     <string name="percent_text_before">Achieve </string>
@@ -50,7 +50,4 @@
     <!--App Bar-->
     <string name="edit_app_bar">Edit</string>
     <string name="delete_app_bar">Delete</string>
-
-    <!--暫定-->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
